Move `CheckPoint` into file, tidy canonical_head
